requirements.txt
Django
djangorestframework==3.3.3
psycopg2
Celery
django-celery-results
lxml
cssselect